Fix bootargs with U-Boot SPL
U-Boot SPL doesn't load the bootargs properly when it's running inside
of this fork of QEMU because because the bootargs struct ends up not
being inside NON_SECURE_SRAM_START and NON_SECURE_SRAM_END. Hack QEMU to
put it in a place where U-Boot is happy with it. This is super duper
hacky but it works.